Move select-points to elements, tool node move gets the event from node-editor
